pyproject.toml
[tool.poetry]
name = "fancy_fashion"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
tensorflow = "^2.8.0"
Pillow = "^9.0.1"
matplotlib = "^3.5.1"
typer = "^0.4.0"
joblib = "^1.1.0"
kfp = "^1.8.11"
google-cloud-aiplatform = "^1.11.0"

[tool.poetry.group.dev.dependencies]
black = "^22.1.0"
pre-commit = "^2.17.0"
ipykernel = "^6.9.2"
pylint = "^2.13.1"
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.'BASIC']
good-names = 'X,x,y,i,ii'

[tool.pylint.'MESSAGES CONTROL']
disable = 'missing-module-docstring,import-error'
